Add core_hours to Tally API
Also moved some TallySnapshot model tests out
of the resource test and into its own test class.
