+++ api/rhsm-subscriptions-api-spec.yaml
@@ -687,6 +687,7 @@ components:
         - Satellite
         - Satellite Capsule
         - Satellite Server
+        - OpenShift-metrics
     Uom:
       type: string
       enum:
@@ -852,6 +853,11 @@ components:
         cores:
           type: integer
           format: int32
+          minimum: 0
+          default: 0
+        core_hours:
+          type: number
+          format: double
           minimum: 0
           default: 0
         sockets:
